updated task design and structure
