+++ icons/IconClock.tsx
@@ -1,0 +1,17 @@
+export const IconClock = ({ size = "2em" }) => {
+  return (
+    <svg
+      width={size}
+      height={size}
+      fill="currentColor"
+      version="1.1"
+      viewBox="0 0 600 550"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g>
+        <path d="m350 87.5c-51.055 0-100.02 20.281-136.12 56.383-36.102 36.102-56.383 85.062-56.383 136.12s20.281 100.02 56.383 136.12c36.102 36.102 85.062 56.383 136.12 56.383s100.02-20.281 136.12-56.383c36.102-36.102 56.383-85.062 56.383-136.12-0.054688-51.039-20.355-99.969-56.441-136.06-36.09-36.086-85.02-56.387-136.06-56.441zm0 350c-41.773 0-81.832-16.594-111.37-46.129-29.535-29.539-46.129-69.598-46.129-111.37s16.594-81.832 46.129-111.37c29.539-29.535 69.598-46.129 111.37-46.129s81.832 16.594 111.37 46.129c29.535 29.539 46.129 69.598 46.129 111.37-0.046875 41.758-16.656 81.793-46.184 111.32-29.523 29.527-69.559 46.137-111.32 46.184z" />
+        <path d="m367.5 272.75v-97.754c0-6.2539-3.3359-12.031-8.75-15.156s-12.086-3.125-17.5 0-8.75 8.9023-8.75 15.156v105c0 4.6406 1.8438 9.0898 5.1289 12.371l52.5 52.5c4.4453 4.293 10.824 5.9258 16.785 4.293 5.957-1.6367 10.613-6.293 12.25-12.25 1.6328-5.9609 0-12.34-4.293-16.785z" />
+      </g>
+    </svg>
+  );
+};
