+++ locales/en.ts
@@ -15,8 +15,10 @@ export const en = {
     tasksList: "Tasks list",
     createTaskMenu: "Create task",
     createTaskTitle: "Create a new task",
-    taskLabelName: "Name",
-    taskLabelDetails: "Details",
+    taskLabelName: "Task",
+    taskPlaceholderName: "Pick up the groceries",
+    taskLabelDetails: "Description",
+    taskPlaceholderDetails: "At the local store, from 6pm on. Bring cash",
     taskLabelDueDate: "Due date",
     taskLabelAssignee: "Assignee",
     createTaskButton: "Create task",
@@ -27,7 +29,7 @@ export const en = {
     errorAssignTask: "Failed to assign task",
     errorDeleteTask: "Failed to delete task",
     errorRescheduleTask: "Failed to reschedule task",
-    dueBy: "Due by",
+    dueBy: "Due",
     taskCompleted: "Completed",
     assignToMe: "Assign to me",
     noAssignee: "No one yet",
