+++ styles/Form.module.scss
@@ -1,66 +1,83 @@
+@use "colors";
+@use "fonts";
 
-.formBlock {
-    margin-bottom: 1rem;
-  }
-  
-  .label {
+.formInputBlock {
+  margin-bottom: 1rem;
+
+  label {
     display: block;
     margin-bottom: 0.2rem;
   }
-  
-  .input{
+
+  textarea,
+  input,
+  select {
     width: 100%;
     padding: 1rem;
-    background: var(--color-pink-light);
-    border: 1px solid black;
+    border: 1px solid colors.$secondary;
     border-radius: 5px;
+    background: colors.$fourth;
     font-family: inherit;
-    font-size: var(--font-size-normal);
-  }
-    .input:disabled{
+    font-size: inherit;
+
+    &:disabled {
       background: var(--color-gray-light);
       cursor: not-allowed;
     }
-  
-  .button{
-    width: 100%;
-    padding: 1rem;
-    color: var(--color-gray-dark);
-    font-size: var(--font-size-normal);
-    font-weight: bold;
-    cursor: pointer;
   }
-    .button:disabled {
-      color: var(--color-gray-light);
-      cursor: not-allowed;
-    }
-  .buttonGreen {
-    background: var(--color-green);
-    border: 1px solid var(--color-green-dark);
+
+  textarea {
+    resize: none; // TODO: see if scrolling works well for description field
   }
-    .buttonGreen:active {
-      background: var(--color-green-dark);
-    }
-    .buttonGreen:disabled {
-      background: var(--color-green-light);
-    }
-  .buttonBlue {
-    background: var(--color-blue);
-    border: 1px solid var(--color-blue-dark);
+}
+
+@mixin button {
+  width: 100%;
+  padding: 1rem;
+  color: colors.$thirdary;
+  border-radius: 5px;
+  font-size: fonts.$bigger;
+  font-weight: bold;
+  cursor: pointer;
+
+  &:disabled {
+    color: colors.$thirdary-light;
+    cursor: not-allowed;
   }
-    .buttonBlue:active {
-      background: var(--color-blue-dark);
-    }
-    .buttonBlue:disabled {
-      background: var(--color-blue-light);
-    }
-  .buttonRed {
-    background: var(--color-red);
-    border: 1px solid var(--color-red-dark);
+}
+.buttonGreen {
+  @include button;
+  background: colors.$green;
+  border: 1px solid colors.$green-dark;
+
+  &:active {
+    background: colors.$green-dark;
   }
-    .buttonRed:active {
-      background: var(--color-red-dark);
-    }
-    .buttonRed:disabled {
-      background: var(--color-red-light);
-    }
+  &:disabled {
+    background: colors.$green-light;
+  }
+}
+.buttonRed {
+  @include button;
+  background: colors.$red;
+  border: 1px solid colors.$red-dark;
+
+  &:active {
+    background: colors.$red-dark;
+  }
+  &:disabled {
+    background: colors.$red-light;
+  }
+}
+.buttonBlue {
+  @include button;
+  background: colors.$blue;
+  border: 1px solid colors.$blue-dark;
+
+  &:active {
+    background: colors.$blue-dark;
+  }
+  &:disabled {
+    background: colors.$blue-light;
+  }
+}
