+++ styles/Tasks.module.scss
@@ -1,36 +1,78 @@
+@use "colors";
+@use "fonts";
+
 .tasksListItem {
-    display: grid;
-    grid-template-columns: 4fr 1fr;
-    margin-bottom: 1rem;
-    width: 100%;
-    padding: 1rem;
-    border: 1px solid var(--color-gray);
-    background: var(--color-gray);
-    cursor: pointer;
+  margin-bottom: 1rem;
+  padding: 0.8rem 0.8rem;
+  color: colors.$thirdary;
+  border: 1px solid colors.$secondary;
+  border-radius: 5px;
+  background: colors.$fourth;
+  cursor: pointer;
+
+  &:hover,
+  &:active {
+    background: colors.$fifth;
+    border: 1px solid colors.$secondary-dark;
   }
-    .tasksListItem:hover{
-      border: 1px solid var(--color-pink-light);
-    }
-  .tasksListName {
-    margin-bottom: 1rem;
+
+  h3 {
+    margin-bottom: 0.3rem;
   }
-  .tasksListDueDate {
-    font-style: italic;
-    font-size: var(--font-size-small);
-  }
-  .tasksListConfirmation {
+
+  .dueDate {
     display: flex;
     align-items: center;
-    justify-content: center;
+    font-size: fonts.$smallest;
+
+    &:not(:last-child) {
+      margin-bottom: 0.5rem;
+    }
+
+    svg {
+      margin-left: -0.5rem;
+      margin-right: 0.2rem;
+    }
   }
-  
-  .taskDetails {
+
+  .description {
+    // TODO: check if cleaner way to do this
+    overflow: hidden;
+    text-overflow: ellipsis;
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    line-clamp: 2;
+    -webkit-box-orient: vertical;
+  }
+}
+
+.taskDetails {
+  margin-bottom: 2rem;
+
+  h2 {
+    margin-bottom: 0.5rem;
+    font-size: fonts.$bigger;
+  }
+
+  p {
+    margin-bottom: 0.5rem;
+  }
+
+  .dueDate {
+    display: flex;
+    align-items: center;
+
+    svg {
+      margin-left: -0.5rem;
+      margin-right: 0.2rem;
+    }
+  }
+}
+
+.taskActions {
+  padding: 0 2rem;
+
+  button {
     margin-bottom: 1rem;
   }
-  .taskActions {
-    padding: 0 2rem;
-  }
-    .taskActions button {
-      margin-bottom: 1rem;
-    }
-  
+}
