+++ styles/_fonts.scss
@@ -1,1 +1,4 @@
+$smallest: 0.8rem;
+$normal: 1rem;
+$bigger: 1.2rem;
 $biggest: 1.4rem;
