+++ styles/globals.scss
@@ -1,30 +1,5 @@
 @use "colors";
 
-:root {
-  --color-gray: #0bbcd6;
-  --color-gray-light: #8d8d99;
-  --color-gray-dark: #0bbcd6;
-
-  --color-pink-medium: #ffe4b5;
-  --color-pink-dark: #e9967a;
-
-  --color-green: #41bf7a;
-  --color-green-light: #81caa2;
-  --color-green-dark: #349961;
-
-  --color-blue: #87cefa;
-  --color-blue-light: #d3ebfb;
-  --color-blue-dark: #6ba4c7;
-
-  --color-red: #ff6347;
-  --color-red-light: #ffa494;
-  --color-red-dark: #cc4f39;
-
-  --font-size-small: 0.8rem;
-  --font-size-normal: 1rem;
-  --font-size-big: 1.2rem;
-}
-
 html,
 body {
   padding: 0;
